requirements.txt
Django==4.0
django_mptt==0.13.4
